todo_project/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.20
django-decouple==2.1
django-rest-authtoken==2.1.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1


# background process and cache
celery
redis
django-celery-beat

# caching
requests